+++ src/lib/Main.cpp
@@ -170,7 +170,7 @@ int Main::setParameter(const Parameter *p, size_t startIndex,
 }
 
 /////////////////////////////////////////////////////////////////////////////
-void Main::processPoll(size_t delay_ms,
+void Main::processPoll(unsigned int delay_ms,
         const PdServ::Signal * const *s, size_t nelem,
         void * const *pollDest, struct timespec *t) const
 {
